Assignments/Prompt_Assignments/assignment_three/requirements.txt
flask
openai
python-dotenv
pandas
flask-socketio
eventlet